requirements.txt
Django==1.6.2
Markdown==2.4
PyStemmer==1.2.0
argparse==1.2.1
beautifulsoup4==4.3.2
cobe==2.1.0
django-filter==0.7
djangorestframework==2.3.13
gunicorn==18.0
python-irclib==0.4.8
wsgiref==0.1.2
